[code-reference/ExternalMergeSort] feat: Merge files

+++ src/code-references/ExternalMergeSort/src/util/FileManager.kt
@@ -9,34 +9,36 @@ import java.io.FileWriter
 import java.util.*
 import kotlin.Comparator
 
-class FileManager(private val partitionConfigurator: MergeSortPartitionConfigurable) {
+interface FileManagerConfigurator {
+    val partition: MergeSortPartitionConfigurable
+    val inputFileName: String
+    val outputFileName: String
+}
 
-    companion object {
-        private val INPUT_FILE_NAME = "input.txt"
-        private val OUTPUT_FILE_NAME = "output.txt"
-    }
+class FileManager(private val configurator: FileManagerConfigurator) {
 
-    private val inputFile = File(INPUT_FILE_NAME)
-    private val outputFile = File(OUTPUT_FILE_NAME)
+    private val inputFile = File(configurator.inputFileName)
+    private val outputFile = File(configurator.outputFileName)
 
     fun generateInput(generator: RandomGenerative) {
-        val size = partitionConfigurator.countPartitions * partitionConfigurator.sizePartition
+        val size = configurator.partition.countPartitions * configurator.partition.sizePartition
         inputFile.printWriter().use { out ->
             (0 until size)
                     .forEach { out.print("${generator.nextInt()} ") }
         }
     }
 
-    fun createInitialRuns(sortDirector: Sortable<Int>) {
+    // Creates sorted output files.
+    fun createScratchFiles(sortDirector: Sortable<Int>) {
         val scanner = Scanner(inputFile)
-        val outputFiles = Array(partitionConfigurator.countPartitions, { File("$it") })
-        val array = Array(partitionConfigurator.sizePartition, { 0 })
+        val outputFiles = Array(configurator.partition.countPartitions, { File("$it") })
+        val array = Array(configurator.partition.sizePartition, { 0 })
 
         var outputFileIndex = 0
 
         while (scanner.hasNextInt()) {
             var i = 0
-            while (scanner.hasNextInt() && i < partitionConfigurator.sizePartition) {
+            while (scanner.hasNextInt() && i < configurator.partition.sizePartition) {
                 array[i++] = scanner.nextInt()
             }
 
@@ -52,47 +54,52 @@ class FileManager(private val partitionConfigurator: MergeSortPartitionConfigura
     }
 
     fun mergeFiles() {
-        outputFile.delete()
-        val inputFiles = Array(partitionConfigurator.countPartitions, { File("$it") })
-        val priorityQueue = PriorityQueue<MinHeapNode>(
-                partitionConfigurator.countPartitions,
+        if (outputFile.exists()) {
+            outputFile.delete()
+        }
+
+        val inputScanners = getInputScanners()
+        val minHeap = PriorityQueue<MinHeapNode>(
+                configurator.partition.countPartitions,
                 Comparator { o1: MinHeapNode, o2: MinHeapNode -> o1.element - o2.element }
         )
 
-        var i = 0
-        while (i < partitionConfigurator.countPartitions) {
-            val scanner = Scanner(inputFiles[i])
+        // Push first element of each file into the min-heap
+        // along with list number (index).
+        for (i in 0 until configurator.partition.countPartitions) {
+            val scanner = inputScanners[i]
             if (scanner.hasNextInt()) {
-                priorityQueue.add(MinHeapNode(scanner.nextInt(), i++))
-            } else {
-                break
+                minHeap.add(MinHeapNode(scanner.nextInt(), i))
             }
         }
 
-        val fileWriter = FileWriter(OUTPUT_FILE_NAME, true)
-        var scanner: Scanner? = null
-        var count = 0
+        val fileWriter = FileWriter(configurator.outputFileName, true)
 
-        while (count < i) {
-            val root = priorityQueue.poll()
-            fileWriter.write("${root.element} ")
+        while (!minHeap.isEmpty()) {
+            val minNode = minHeap.poll()
+            fileWriter.write("${minNode.element} ")
 
-            if (scanner == null) {
-                scanner = Scanner(inputFiles[root.index])
-            }
-
+            // Take next element from the same file and
+            // insert it into min-heap.
+            val scanner = inputScanners[minNode.index]
             if (scanner.hasNextInt()) {
-                root.element = scanner.nextInt()
-            } else {
-                scanner.close()
-                scanner = null
-                root.element = Int.MAX_VALUE
-                count++
+                minNode.element = scanner.nextInt()
+                minHeap.add(minNode)
             }
-
-            priorityQueue.add(root)
         }
 
         fileWriter.close()
+        closeScanners(inputScanners)
+    }
+
+    private fun getInputScanners(): Array<Scanner> {
+        val inputFiles = Array(configurator.partition.countPartitions, { File("$it") })
+        return Array(inputFiles.size, { Scanner(inputFiles[it]) })
+    }
+
+    private fun closeScanners(scanners: Array<Scanner>) {
+        for (anScanner in scanners) {
+            anScanner.close()
+        }
     }
 }
